% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_copilot_whitelist.R
\name{default_allowlist}
\alias{default_allowlist}
\title{Curated R and GitHub domains}
\usage{
default_allowlist()
}
\value{
A character vector of suggested domains.
}
\description{
A curated set of R and GitHub domains to add to the GitHub Copilot coding
agent firewall allowlist, or other agent domain whitelists.
}
\examples{
default_allowlist()
}
