% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_skill_create_issue.R
\name{.bug_reports_from_remote}
\alias{.bug_reports_from_remote}
\title{Update BugReports from repo info}
\usage{
.bug_reports_from_remote(call = caller_env())
}
\arguments{
\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
(\code{character(1)}) The BugReports URL.
}
\description{
Update BugReports from repo info
}
\keyword{internal}
