% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.call_gh}
\alias{.call_gh}
\title{Call the GitHub API}
\usage{
.call_gh(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[gh:gh]{gh::gh()}}.}
}
\value{
The API response.
}
\description{
Thin wrapper around \code{\link[gh:gh]{gh::gh()}} to facilitate mocking in tests.
}
\keyword{internal}
