% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_file_exists}
\alias{.check_file_exists}
\title{Check whether a file exists and act on it}
\usage{
.check_file_exists(path, overwrite, call = rlang::caller_env())
}
\arguments{
\item{overwrite}{(\code{logical(1)}) Whether to overwrite an existing file.
Defaults to \code{TRUE}.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Deletes the file if it exists and \code{overwrite = TRUE}. Errors if the file
exists and \code{overwrite = FALSE}.
}
\keyword{internal}
