% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{.check_path_writable}
\alias{.check_path_writable}
\title{Check whether a path is writable}
\usage{
.check_path_writable(path, overwrite, call = caller_env())
}
\arguments{
\item{path}{(\code{character(1)}) Absolute path to the file.}

\item{overwrite}{(\code{logical(1)}) Whether to overwrite existing files. Defaults
to \code{FALSE}.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
\code{NULL}, invisibly. In part called for side effects: Deletes the file
if it exists and \code{overwrite = TRUE}. Errors if the file exists and
\code{overwrite = FALSE}.
}
\description{
Check whether a path is writable
}
\keyword{internal}
