% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_skill_create_issue.R
\name{.get_create_issue_data}
\alias{.get_create_issue_data}
\title{Collect template data for the create-issue skill}
\usage{
.get_create_issue_data(gh_token, call = caller_env())
}
\arguments{
\item{gh_token}{(\code{character(1)}) A GitHub personal access token. Defaults to
\code{gh::gh_token()}.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
(\code{list}) Named list of whisker template variables.
}
\description{
Collect template data for the create-issue skill
}
\keyword{internal}
