% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_desc_fields}
\alias{.get_desc_fields}
\title{Get non-NA fields from DESCRIPTION}
\usage{
.get_desc_fields(fields, call = caller_env())
}
\arguments{
\item{fields}{(\code{character}) Field name(s) to read from \code{DESCRIPTION}.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
(\code{list}) Named list of non-\code{NA} field values from \code{DESCRIPTION}.
}
\description{
Get non-NA fields from DESCRIPTION
}
\keyword{internal}
