% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{.lines_insert_after}
\alias{.lines_insert_after}
\title{Insert lines into a character vector after a given index}
\usage{
.lines_insert_after(lines, idx, new_lines)
}
\arguments{
\item{lines}{(\code{character}) Lines of a file, as returned by \code{\link[=readLines]{readLines()}}.}

\item{idx}{(\code{integer(1)}) Position after which \code{new_lines} will be inserted.}

\item{new_lines}{(\code{character}) Lines to insert.}
}
\value{
(\code{character}) Updated lines with \code{new_lines} spliced in after \code{idx}.
}
\description{
Insert lines into a character vector after a given index
}
\keyword{internal}
