% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{.path_pkg}
\alias{.path_pkg}
\title{Build a path to a inst file within installed or dev pkgskills}
\usage{
.path_pkg(...)
}
\arguments{
\item{...}{Path components, passed to \code{\link[fs:path_package]{fs::path_package()}}.}
}
\value{
(\code{character(1)}) Absolute path within the pkgskills package.
}
\description{
Build a path to a inst file within installed or dev pkgskills
}
\keyword{internal}
