% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{.path_proj_save_as}
\alias{.path_proj_save_as}
\title{Build and validate a project-relative output path}
\usage{
.path_proj_save_as(save_as, overwrite, call = caller_env())
}
\arguments{
\item{save_as}{(\code{character(1)}) Output file path, relative to the project
root.}

\item{overwrite}{(\code{logical(1)}) Whether to overwrite an existing file.
Defaults to \code{TRUE}.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
(\code{character(1)}) Absolute path to \code{save_as} within the active
project.
}
\description{
Build and validate a project-relative output path
}
\keyword{internal}
