% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.path_template}
\alias{.path_template}
\title{Build a path to a file in \verb{inst/templates/}}
\usage{
.path_template(...)
}
\arguments{
\item{...}{Path components appended after \code{"templates"}, passed to
\code{\link[=.path_pkg]{.path_pkg()}}.}
}
\value{
(\code{character(1)}) Absolute path to the template file.
}
\description{
Build a path to a file in \verb{inst/templates/}
}
\keyword{internal}
