% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{.pkg_abort}
\alias{.pkg_abort}
\title{Raise a package-scoped error}
\usage{
.pkg_abort(message, subclass, call = rlang::caller_env())
}
\arguments{
\item{message}{(\code{character(1)}) A \code{{cli}} message string.}

\item{subclass}{(\code{character(1)}) Error subclass.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
Does not return.
}
\description{
Raise a package-scoped error
}
\keyword{internal}
