% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_string}
\alias{.to_string}
\title{Coerce to a non-null, non-empty character scalar}
\usage{
.to_string(x, x_arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The value to coerce.}

\item{x_arg}{(\code{character(1)}) Argument name for \code{x}, used in error messages.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
(\code{character(1)}) \code{x} coerced to a character scalar.
}
\description{
Coerce to a non-null, non-empty character scalar
}
\keyword{internal}
