% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_skill_r_code.R
\name{.use_conditions}
\alias{.use_conditions}
\title{Install the conditions template}
\usage{
.use_conditions(overwrite = FALSE, open = rlang::is_interactive())
}
\arguments{
\item{overwrite}{(\code{logical(1)}) Whether to overwrite existing file(s).
Defaults to \code{FALSE}.}

\item{open}{(\code{logical(1)}) Whether to open the file after creation.}
}
\value{
The path to the installed conditions file, invisibly.
}
\description{
Install the conditions template
}
\keyword{internal}
