% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.use_template}
\alias{.use_template}
\title{Wrapper around \code{\link[usethis:use_template]{usethis::use_template()}}}
\usage{
.use_template(
  template,
  save_as,
  data = list(),
  open = FALSE,
  call = caller_env()
)
}
\arguments{
\item{template}{(\code{character(1)}) Template name within \verb{inst/templates/}.}

\item{save_as}{(\code{character(1)}) Output file path, relative to the project
root.}

\item{data}{(\code{list}) Named list of whisker template variables for rendering.}

\item{open}{(\code{logical(1)}) Whether to open the file after creation.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
Called for side effects.
}
\description{
Wrapper around \code{\link[usethis:use_template]{usethis::use_template()}}
}
\keyword{internal}
