% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgskills-package.R
\docType{package}
\name{pkgskills-package}
\alias{pkgskills}
\alias{pkgskills-package}
\title{pkgskills: AI Skills for R Package Development}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of curated, opinionated skills and agent instructions to improve agentic coding of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/api2r/pkgskills}
  \item \url{https://api2r.github.io/pkgskills/}
  \item Report bugs at \url{https://github.com/api2r/pkgskills/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
}

}
\keyword{internal}
