% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params}
\alias{.shared-params}
\title{Shared parameters}
\arguments{
\item{call}{(\code{environment}) The caller environment for error messages.}

\item{open}{(\code{logical(1)}) Whether to open the file after creation.}

\item{save_as}{(\code{character(1)}) Output file path, relative to project root.}

\item{x_arg}{(\code{character(1)}) Argument name for \code{x}, used in error messages.}
}
\description{
These parameters are used in multiple functions. They are defined here to
make them easier to import and to find.
}
\keyword{internal}
