% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.use_template_as_is}
\alias{.use_template_as_is}
\title{Write a template verbatim to a project file}
\usage{
.use_template_as_is(
  template,
  save_as,
  overwrite = FALSE,
  open = FALSE,
  call = caller_env()
)
}
\arguments{
\item{save_as}{(\code{character(1)}) Output file path, relative to the project
root.}

\item{overwrite}{(\code{logical(1)}) Whether to overwrite an existing file.
Defaults to \code{FALSE}.}

\item{open}{(\code{logical(1)}) Whether to open the file after creation.}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
Called for side effects.
}
\description{
Write a template verbatim to a project file
}
\keyword{internal}
